Move TransactionStorage to trie package
